simplify and format
